+++ packages/ts-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @asyncui/ts-config
 
+## 2.0.2
+
+### Patch Changes
+
+- fix Alert component
+
 ## 2.0.1
 
 ### Patch Changes